MD20250105-02 Replacing length with p_length
MD20250105-02
In function addLength, in snake_helpers.c and
as an example, and to differentiate where we are using pointers to an int (to length) as opposed the an int (x or y maybe)
I have introduced p_ at the start of the variable name for the pointer to the length.
We don't have to do thins now for all parameters that are passed as pointer to functions, but this technique may help in distinguishing int from pointers to int.