MD20250105-05
Changing i and j to row and column.
There is no god argument in favour of using more obscure variable names like i and j when more descriptive ones could do the same job.
Again, this is not a right/wrong situation, but is a matter of style. It makes the code easier to read.
Some programmers, particularly young ones, prefer to use obscure variable names because they think it makes them look smarter.